+++ src/debugger/calva-debug.ts
@@ -197,7 +197,7 @@ class CalvaDebugSession extends LoggingDebugSession {
 
         // Pass scheme in path argument to Source contructor so that if it's a jar file it's handled correctly
         const source = new Source(basename(debugResponse.file), debugResponse.file);
-        const name = tokenCursor.getFunction();
+        const name = tokenCursor.getFunctionName();
         const stackFrames = [new StackFrame(0, name, source, line + 1, column + 1)];
 
         response.body = {