+++ src/status.ts
@@ -1,11 +1,25 @@
+import * as vscode from 'vscode';
 import * as namespace from './namespace';
 import statusbar from './statusbar';
+import * as state from './state';
 
-function update() {
+function updateNeedReplUi(isNeeded: boolean, context = state.extensionContext) {
+    context.workspaceState.update('needReplUi', isNeeded);
+    update(context);
+}
+
+function shouldshowReplUi(context = state.extensionContext): boolean {
+    return context.workspaceState.get('needReplUi') || !state.config().hideReplUi;
+}
+
+function update(context = state.extensionContext) {
+    vscode.commands.executeCommand('setContext', 'calva:showReplUi', shouldshowReplUi(context));
     namespace.updateREPLSessionType();
-    statusbar.update();
+    statusbar.update(context);
 }
 
 export default {
-    update
+    update,
+    updateNeedReplUi,
+    shouldshowReplUi
 };